+++ src/main/java/com/rocket/rocket/domain/ExDto.java
@@ -1,4 +1,4 @@
-package com.rocket.rocket.dto;
+package com.rocket.rocket.domain;
 
 import lombok.Getter;
 import lombok.Setter;
