+++ src/main/java/com/rocket/rocket/domain/MakeDto.java
@@ -1,15 +1,13 @@
-package com.rocket.rocket.dto;
+package com.rocket.rocket.domain;
 
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
-public class TakeDto {
+public class MakeDto {
 
-	private String take_num;
-	
-	private String regDate;
+	private String make_num;
 	
 	private String user_num;
