+++ src/main/java/com/rocket/rocket/domain/RoleVO.java
@@ -1,0 +1,11 @@
+package com.rocket.rocket.domain;
+
+import lombok.Data;
+
+@Data
+public class RoleVO {
+
+	private byte role_num;
+	
+	private String roleName;
+}
