+++ src/main/java/com/rocket/rocket/domain/ScoreVO.java
@@ -1,0 +1,9 @@
+package com.rocket.rocket.domain;
+
+import lombok.Data;
+
+@Data
+public class ScoreVO {
+
+	private int lec_sc_star;
+}
