+++ src/main/java/com/rocket/rocket/domain/SubVO.java
@@ -1,0 +1,13 @@
+package com.rocket.rocket.domain;
+
+import lombok.Data;
+
+@Data
+public class SubVO {
+
+	private String sub_num;
+	
+	private Long price;
+	
+	private String type;
+}
