+++ src/main/java/com/rocket/rocket/domain/UserCtgyVO.java
@@ -1,0 +1,13 @@
+package com.rocket.rocket.domain;
+
+import lombok.Data;
+
+@Data
+public class UserCtgyVO {
+
+	private String user_ctgy_num;
+	
+	private String ctgy_title;
+	
+	private String user_num;
+}
