+++ src/main/java/com/rocket/rocket/domain/UserRoleVO.java
@@ -1,0 +1,12 @@
+package com.rocket.rocket.domain;
+
+import lombok.Data;
+
+@Data
+public class UserRoleVO {
+	
+	private String user_num;
+	
+	private byte user_role;
+
+}
