--- src/main/java/com/rocket/rocket/dto/MakeDto.java
@@ -1,15 +1,0 @@
-package com.rocket.rocket.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class MakeDto {
-
-	private String make_num;
-	
-	private String user_num;
-	
-	private String class_num;
-}
