--- src/main/java/com/rocket/rocket/dto/ReplyDto.java
@@ -1,23 +1,0 @@
-package com.rocket.rocket.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class ReplyDto {
-
-	private String reply_num;
-	
-	private String content;
-	
-	private String nickname;
-	
-	private String regDate;
-	
-	private String modiDate;
-	
-	private String delDate;
-	
-	private String comm_num;
-}
