--- src/main/java/com/rocket/rocket/dto/RoleDto.java
@@ -1,13 +1,0 @@
-package com.rocket.rocket.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class RoleDto {
-
-	private byte role_num;
-	
-	private String name;
-}
