--- src/main/java/com/rocket/rocket/dto/ScoreDto.java
@@ -1,11 +1,0 @@
-package com.rocket.rocket.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class ScoreDto {
-
-	private int lec_sc_star;
-}
