--- src/main/java/com/rocket/rocket/dto/SubDto.java
@@ -1,15 +1,0 @@
-package com.rocket.rocket.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class SubDto {
-
-	private String sub_num;
-	
-	private Long price;
-	
-	private String type;
-}
