--- src/main/java/com/rocket/rocket/dto/UserCtgyDto.java
@@ -1,15 +1,0 @@
-package com.rocket.rocket.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class UserCtgyDto {
-
-	private String user_ctgy_num;
-	
-	private String ctgy_title;
-	
-	private String user_num;
-}
